use createStorageSpace generator in tests
