+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/generators/BagRequestGenerators.scala
@@ -1,23 +1,16 @@
 package uk.ac.wellcome.platform.archive.common.generators
 
 import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
+import uk.ac.wellcome.platform.archive.common.models.BagRequest
 import uk.ac.wellcome.platform.archive.common.models.bagit.BagLocation
-import uk.ac.wellcome.platform.archive.common.models.{BagRequest, StorageSpace}
 
-trait BagRequestGenerators extends RandomThings {
-  def createBagRequest() = createBagRequestWith()
+trait BagRequestGenerators extends RandomThings with BagLocationGenerators {
+  def createBagRequest(): BagRequest = createBagRequestWith()
 
   def createBagRequestWith(
-    bagLocation: BagLocation = createBagLocation): BagRequest =
+    bagLocation: BagLocation = createBagLocation()): BagRequest =
     BagRequest(
       requestId = randomUUID,
       bagLocation = bagLocation
     )
-
-  def createBagLocation() =
-    BagLocation(
-      storageNamespace = randomAlphanumeric(),
-      storagePrefix = Some(randomAlphanumeric()),
-      storageSpace = StorageSpace(randomAlphanumeric()),
-      bagPath = randomBagPath)
 }
